Enable conversions between formats
These changes enable 12 of the 13 formats to be converted between them. Local transactions
have voluntarily been excluded as they require an lgsl code being inputed before creation - this would
be considerable work for little benefit, as formats are rarely converted to a local transaction.

The tests cover all conversion cases, except when converting from programme to guide editions - editions
with parts are a special case that will be dealt with in a separate commit.