Convert from non-parted to parted edition
This tests for non-parted documents to be converted into parted documents
such as programme editions and guide editions. When converting, the whole_body
of the source will be copied to a new part of the target class.

If the target class also has default parts, they will be created in addition to
the part receiving the whole_body.

The code to make this test pass is in govuk_content_models.