The Isles of Scilly are a unitary authority
The Isles of Scilly have a unique area type in MapIt, COI. It's an unusual
local authority but is most similar to a unitary authority - Cornwall doesn't
cover it so this is the only local authority which covers the islands.

Frontend already treats COI as unitary, so users get redirected to the right
page for local transactions, but the results page for district-level local
transactions is incorrectly rendered at the moment. This is because Publisher
has been treating the LA as a county on import, so when content API gets a
request for a district-level local transaction with the SNAC for this LA, it
can't find a district or unitary authority with that SNAC in its database so
returns the artefact with no LA details. Frontend then interprets that as
indicating that it's rendering the front page rather than an LA-specific page,
so shows the postcode form again instead of a link (or a message saying we
don't have a link).

Adding COI to the list of area types to treat as unitary fixes this, so that
all local transactions should work for the Isles of Scilly.