WiP: Try sending LGIL and LGSL to publishing API
To unblock lots of things for local links manager we need access to the list
of published (+ draft?) local transactions including their LGSL and LGIL codes.
They're in the publishing API as placeholders already but without those codes.
We would like to be able to get this info from the publishing API to use in
local links manager.

This is to test including the codes to see how it works. It looks like pub API
schema validation does nothing if the schema name is placeholder_* but we
should probably add a schema for placeholder_local_transaction for safety and
also to document that we expect these fields to be there.

This also needs tests...